rename buildspecs to ecs dir and add ecs plan task script
