+++ ecs/create_deploy_stack/sql/utils.sql
