+++ ecs/img/Dockerfile
