+++ ecs/img/install.sh
