+++ ecs/terra_run/update_new_resources.py
